Loop Retropolate_ExchangeRate_PPP
Previously, the script import exchange-rate.do was pulling data from the files gdp.data and retropolate-gdp.dta, which were generated later in the  execution order of main.do. To address this issue, 1. a table snapshot named SUYU-retropolate-gdp-output.dta was created by replicating the retropolate-gdp table. 2. This snapshot is used to feed the process in the import-exchange-rate code. Additionally, in the main.do file, the order of table production was modified so that GDP is calculated immediately after the price index. Now, 3. The PPP then use this new order to obtain the values of KP and YU from the gap.dta file.