Correction of the Price Index for RU
By dropping the Price index provided by Novokmet for Russia (def_east), the Program selects the index from Novokmet, Piketty & Zucman (2018)(def_wid), which corrects the unusual trends in macro aggregates in the 70's in comparison with the GB and the big fall in the 90's. As so, the hand-made adjustments for RU's index ( from NPZ 2018) were not necessary, so they were deleted.