Complete populations series before 1950 and DE (1937-1944)
Previously, the unusual trends in the "DE" population from 1937 to 1944 were excluded to avoid unnecessary noise. However, the corrections made in the series from Federico-Tena now allow for the recovery of this data. Additionally, the breakdown by gender and age, already calculated for the period before 1950, makes it possible to reimport the series 997 (f, i, and m).